+++ compiler/src/handlers/VariableDeclaration.ts
@@ -1,5 +1,6 @@
 import {
   DestructuringValue,
+  LazyValue,
   LiteralValue,
   ObjectValue,
   SenseableValue,
@@ -66,6 +67,8 @@ const Declare: TDeclareHandler<es.LVal> = (c, scope, node, kind) => {
         return DeclareArrayPattern(c, scope, node, kind);
       case "ObjectPattern":
         return DeclareObjectPattern(c, scope, node, kind);
+      case "AssignmentPattern":
+        return DeclareAssignmentPattern(c, scope, node, kind);
       default:
         throw new CompilerError(
           `Unsupported declaration type: ${node.type}`,
@@ -82,8 +85,7 @@ const DeclareIdentifier: TDeclareHandler<es.Identifier> = (
 ) => {
   const { name: identifier } = node;
   const name = nodeName(node, !c.compactNames && identifier);
-  const out = SenseableValue.from(
-    scope,
+  const out = new SenseableValue(
     name,
     kind === "const" ? EMutability.init : EMutability.mutable
   );
@@ -139,22 +141,11 @@ const DeclareArrayPattern: TDeclareHandler<es.ArrayPattern> = (
 
     members.set(new LiteralValue(i), {
       value: value.out,
-      handler(get) {
-        return c.handle(scope, element, () => {
-          let [init, initInst]: TValueInstructions<IValue | null> = [null, []];
-
-          try {
-            [init, initInst] = get();
-          } catch (e) {
-            // catches the "cannot get undefined member" error
-            // TODO: refactor and add error codes
-            // so we don't have to do string checks
-            if (
-              !(e instanceof CompilerError) ||
-              !e.message.includes("is not present in")
-            )
-              throw e;
-          }
+      handler(get, propExists, scope) {
+        const [init, inst] = c.handle(scope, element, () => {
+          let [init, initInst] = propExists()
+            ? get()
+            : [value.defaultInit ? new LiteralValue(null) : null, []];
 
           if (!init)
             throw new CompilerError(
@@ -162,8 +153,16 @@ const DeclareArrayPattern: TDeclareHandler<es.ArrayPattern> = (
               element
             );
 
-          return value.handler(init, initInst);
+          const getDefault = value.defaultInit;
+          if (getDefault) {
+            init = pipeInsts(
+              init["??"](scope, new LazyValue(() => getDefault()), value.out),
+              initInst
+            );
+          }
+          return [init, initInst];
         });
+        return value.handler(init, inst);
       },
     });
   }
@@ -213,15 +212,29 @@ const DeclareObjectPattern: TDeclareHandler<es.ObjectPattern> = (
 
     members.set(keyInst[0], {
       value: value.out,
-      handler(get) {
-        return c.handle(scope, propValue, () => {
-          const [init, initInst] = get();
-          return value.handler(init, initInst);
+      handler(get, propExists, scope) {
+        const [init, inst] = c.handle(scope, prop, () => {
+          let [init, initInst] =
+            propExists() || !value.defaultInit
+              ? get()
+              : [new LiteralValue(null), []];
+
+          const getDefault = value.defaultInit;
+          if (getDefault) {
+            init = pipeInsts(
+              init["??"](scope, new LazyValue(() => getDefault()), value.out),
+              initInst
+            );
+          }
+
+          return [init, [...keyInst[1], ...initInst]];
         });
+
+        return value.handler(init, inst);
       },
     });
 
-    return [null, [...keyInst[1], ...valueInst]];
+    return [null, valueInst];
   });
 
   const out = new DestructuringValue(members);
@@ -245,6 +258,18 @@ const DeclareObjectPattern: TDeclareHandler<es.ObjectPattern> = (
   return [declarationValue, propertiesInst[1]];
 };
 
+const DeclareAssignmentPattern: TDeclareHandler<es.AssignmentPattern> = (
+  c,
+  scope,
+  node,
+  kind
+) => {
+  const [value, inst] = Declare(c, scope, node.left, kind);
+
+  value.defaultInit = () => c.handleEval(scope, node.right, value.out);
+
+  return [value, inst];
+};
 class DeclarationValue extends VoidValue {
   c: Compiler;
   scope: IScope;
@@ -252,6 +277,7 @@ class DeclarationValue extends VoidValue {
   out: IValue;
   /** Handles the init value */
   handle: DeclarationValue["handler"];
+  defaultInit?: () => TValueInstructions;
 
   constructor({
     c,