+++ compiler/src/values/index.ts
@@ -6,3 +6,4 @@ export * from "./ObjectValue";
 export * from "./SenseableValue";
 export * from "./DestructuringValue";
 export * from "./LazyValue";
+export * from "./AssignmentValue";